Database

Documentation

migrateAll :: Migration

insertBuild :: MonadIO m => Build -> SqlPersistT m ()

Insert a build into the database

getBuildByUUID :: MonadIO m => Text -> SqlPersistT m (Maybe (Entity Build))

Get a build from the database by its unique commit hash

getBuilds :: MonadIO m => SqlPersistT m [Value Text]

Get all builds from the database

data Build

Defines the schema for Build entity with a unique identifier and build details.

Constructors

Build

Fields

buildUuid :: !Text

buildCommit :: !Text

buildStatus :: !Text

buildDate :: !Text

buildLog :: !Text